requirements.txt
annotated-types==0.7.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
configmypy==0.2.0
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
filelock==3.20.2
fonttools==4.61.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
h5py==3.15.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
neuraloperator==2.0.0
numpy==2.2.6
# nvidia-cublas-cu12==12.8.4.1
# nvidia-cuda-cupti-cu12==12.8.90
# nvidia-cuda-nvrtc-cu12==12.8.93
# nvidia-cuda-runtime-cu12==12.8.90
# nvidia-cudnn-cu12==9.10.2.21
# nvidia-cufft-cu12==11.3.3.83
# nvidia-cufile-cu12==1.13.1.3
# nvidia-curand-cu12==10.3.9.90
# nvidia-cusolver-cu12==11.7.3.90
# nvidia-cusparse-cu12==12.5.8.93
# nvidia-cusparselt-cu12==0.7.1
# nvidia-nccl-cu12==2.27.5
# nvidia-nvjitlink-cu12==12.8.93
# nvidia-nvshmem-cu12==3.3.20
# nvidia-nvtx-cu12==12.8.90 commented out for windows
opt_einsum==3.4.0
packaging==25.0
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
protobuf==6.33.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
pytest==9.0.2
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
ruamel.yaml==0.19.1
scipy==1.15.3
sentry-sdk==2.48.0
six==1.17.0
smmap==5.0.2
# sympy==1.14.0
sympy==1.13.1 # changed for windows
tensorly==0.9.0
tensorly-torch==0.5.0
tomli==2.3.0
# torch==2.9.1 commented out for windows
# torchaudio==2.9.1
# torchvision==0.24.1
# triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
wandb==0.23.1
zencfg==0.6.0
